# Conflicts:
#	src/main/java/ivan/vatlin/configuration/DatabaseConfiguration.java
#	src/main/java/ivan/vatlin/controllers/CabinetController.java
#	src/main/java/ivan/vatlin/controllers/CarController.java
#	src/main/java/ivan/vatlin/controllers/OrderController.java
#	src/main/java/ivan/vatlin/controllers/RegistrationController.java
#	src/main/java/ivan/vatlin/controllers/RestApiController.java
#	src/main/java/ivan/vatlin/controllers/UserController.java
#	src/main/java/ivan/vatlin/dao/CarDao.java
#	src/main/java/ivan/vatlin/dao/CarSpecificationDao.java
#	src/main/java/ivan/vatlin/dao/IUserDao.java
#	src/main/java/ivan/vatlin/dao/OrderDao.java
#	src/main/java/ivan/vatlin/dao/UserDao.java
#	src/main/java/ivan/vatlin/dto/Car.java
#	src/main/java/ivan/vatlin/dto/User.java
#	src/main/java/ivan/vatlin/repositories/CarRepository.java
#	src/main/java/ivan/vatlin/repositories/OrderRepository.java
#	src/main/java/ivan/vatlin/repositories/UserRepository.java
#	src/main/java/ivan/vatlin/services/CarBaseService.java
#	src/main/java/ivan/vatlin/services/CarRentService.java
#	src/main/java/ivan/vatlin/services/CarService.java
#	src/main/java/ivan/vatlin/services/CarSpecificationBaseService.java
#	src/main/java/ivan/vatlin/services/OrderBaseService.java
#	src/main/java/ivan/vatlin/services/OrderRentService.java
#	src/main/java/ivan/vatlin/services/OrderService.java
#	src/main/java/ivan/vatlin/services/UserBaseService.java
#	src/main/java/ivan/vatlin/services/UserRentService.java
#	src/main/resources/db.properties

+++ src/main/java/ivan/vatlin/services/CarBaseService.java
@@ -3,13 +3,14 @@ package ivan.vatlin.services;
 import ivan.vatlin.dao.CarDao;
 import ivan.vatlin.dto.Car;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Lazy;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 @Service
-public class CarBaseService implements CarService{
+public class CarBaseService implements CarService {
     @Autowired
     private CarDao carDao;
 
@@ -39,7 +40,7 @@ public class CarBaseService implements CarService{
     }
 
     @Override
-    public int getNumberOfCars() {
+    public long getNumberOfCars() {
         return carDao.getNumberOfCars();
     }
 }
